Move aws_event_stream to smithy_aws_event_stream
