+++ packages/smithy-aws-event-stream/CHANGES.md
@@ -1,0 +1,4 @@
+## Unreleased
+### Features
+* Added basic support for the [Amazon Event Stream](https://smithy.io/2.0/aws/amazon-eventstream.html) specification.
+* Added support for unidirectdional and bidirectional event streams.
