+++ packages/smithy-aws-event-stream/NOTICE
