+++ packages/smithy-aws-event-stream/src/smithy_aws_event_stream/__init__.py
@@ -3,4 +3,4 @@
 
 import importlib.metadata
 
-__version__: str = importlib.metadata.version("aws-event-stream")
+__version__: str = importlib.metadata.version("smithy-aws-event-stream")
