+++ packages/smithy-aws-event-stream/src/smithy_aws_event_stream/events.py
