+++ pyproject.toml
@@ -28,7 +28,7 @@ smithy_core = { workspace = true }
 smithy_http = { workspace = true }
 smithy_json = { workspace = true }
 smithy_aws_core = { workspace = true }
-aws_event_stream = { workspace = true }
+smithy_aws_event_stream = { workspace = true }
 aws_sdk_signers = {workspace = true }
 
 [tool.pyright]
